\docType{data}
\name{gryphon}
\alias{gryphon}
\title{Gryphon pedigree and birth data}
\format{Data frame with parent information and birth data.}
\description{
  This table contains the gryphon data from the tutorial files from
  Wilson et al. 2010 (Wilson, Alastair J., Denis Reale, Michelle
  N. Clements, Michael M. Morrissey, Erik Postma, Craig A. Walling,
  Loeske E. B. Kruuk, and Daniel H. Nussey. "An Ecologist’s Guide to the
  Animal Model." Journal of Animal Ecology 79, no. 1 (January 2010):
  13–26. doi:10.1111/j.1365-2656.2009.01639.x).

  It merges the pedigree data (sire and dam column) and the birth data
  (birth year, sex and birth weight) from the \code{gryphon.ped_WOMBAT}
  and \code{gryphon_uni.dat} files.
}
\keyword{dataset}
\keyword{datasets}
