% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wombat.parse.permutations.h2}
\alias{wombat.parse.permutations.h2}
\title{Parse the results of a permutation test analysis}
\usage{
wombat.parse.permutations.h2(results, na.rm = F)
}
\arguments{
\item{results}{output from \code{\link{wombat.permutation.test.h2}}
  runs be removed? Default \code{FALSE}.}
}
\value{
A vector with the heritability values for each permuted dataset
  (including NAs if \code{na.rm=F}).
}
\description{
Parse the output of \code{\link{wombat.permutation.test.h2}}.
}
\examples{
# use the gryphon dataset from Wilson et al. 2010
#
head(gryphon)

# permutations (20 perms)
b = wombat.permutation.test.h2(data = gryphon, trait.id = "birth.weight",
       animal.id = "animal", father.id = "sire", mother.id = "dam",
       var.animal = 1, var.residual = 1,
       maternal.effect = F, convert.pedigree = F,
       n.perm = 20)
c = wombat.parse.permutations.h2(b)
c
}

