% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wombat.parse.result}
\alias{wombat.parse.result}
\title{Parse the output file from a Wombat run (SumEstimates.out)}
\usage{
wombat.parse.result()
}
\value{
A \code{list} containing the output results.
}
\description{
Parse the information contained in SumEstimates.out after a
  Wombat run and make it available in a R object.
}
\details{
None.
}
\examples{
# generate a random pedigree structure for 20 offsprings, using 5 different
# sires and 7 different dams
set.seed(4)
animal_id = 1:20
sire_id = sample(1:5, size = 20, replace = T)
dam_id = sample(1:7, size = 20, replace = T)

# generate some random traits
weight = rnorm(n = 20, mean = 50)
length = rnorm(n = 20, mean = 100)

# assemble the information into one data frame
ped_pheno_data = data.frame(animal_id, sire_id, dam_id, weight, length)
ped_pheno_data

# save the files
wombat.write.pedigree(data = ped_pheno_data, animal.id = "animal_id",
  father.id = "sire_id", mother.id = "dam_id", convert.pedigree = T,
  file.path = "toto.ped")
wombat.write.data(data = ped_pheno_data, trait = "weight",
  animal.id = "animal_id", father.id = "sire_id", mother.id = "dam_id",
  write.mother = T, trait.factor = 2, convert.pedigree = T,
  filter.unknown.mother = T, file.path = "toto.dat")
wombat.write.parameters(trait = "weight", pedigree.file = "toto.ped",
  data.file = "toto.dat", var.residual = 2, var.animal = 2,
  var.mother = 2, maternal.effect = T, file.path = "toto.par")

# run the analysis
wombat.unit.run(parameter.file = "toto.par")

# parse the results
a = wombat.parse.result()
}

