% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wombat.parse.sensitivity.h2}
\alias{wombat.parse.sensitivity.h2}
\title{Parse the results of a sensitivity analysis}
\usage{
wombat.parse.sensitivity.h2(results)
}
\arguments{
\item{results}{output from \code{\link{wombat.sensitivity.h2}}}
}
\value{
A data frame with two columns: the starting variance values and
  estimated h2 values
}
\description{
Parse the output of \code{\link{wombat.sensitivity.h2}}.
}
\examples{
# use the gryphon dataset from Wilson et al. 2010
#
head(gryphon)
# sensitivity test
var.obs = var(gryphon[["birth.weight"]], na.rm = T)
var.obs
var.animal = seq(1, 7, by = 1)
var.residual = var.obs - var.animal
b = wombat.sensitivity.h2(data = gryphon, trait.id = "birth.weight",
       animal.id = "animal", father.id = "sire", mother.id = "dam",
       var.animal = var.animal, var.residual = var.residual,
       maternal.effect = F, convert.pedigree = F)
c = wombat.parse.sensitivity.h2(b)
c
}

