% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wombat.split.output.line}
\alias{wombat.split.output.line}
\title{Convert a string from Wombat output to a vector}
\usage{
wombat.split.output.line(string)
}
\arguments{
\item{string}{a \code{string}}
}
\value{
A vector of substrings corresponding to the initial string splitted
  along the blank spaces.
}
\description{
Split a string from a Wombat output file into substrings based
  on the blank spaces.
}
\details{
None.
}
\examples{
# example of Wombat output
output = "***** Estimates of residual covariances   ************************************\\n
  Order of fit         =       1\\n
  Covariance matrix\\n
    3.4074\\n
  Matrix of correlations and variance ratios\\n
    0.4725\\n
  Covariances & correlations with approximate sampling errors\\n
    1  COVS Z 1 1        3.40737       0.549224       vrat     0.472   0.082"

lines = strsplit(output, "\\n")[[1]]
for (i in lines) {
  print(wombat.split.output.line(i))
 }
}

