% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wombat.unit.run}
\alias{wombat.unit.run}
\title{Run one Wombat analysis}
\usage{
wombat.unit.run(parameter.file,
  wombat.bin = "/home/matthieu/bin/wombat/wombat")
}
\arguments{
\item{parameter.file}{path to the parameter file (e.g. file generated by
\code{\link{wombat.write.parameters}}). This path MUST end with
\code{".par"}.}

\item{wombat.bin}{path to the wombat executable}
}
\value{
None.
}
\description{
Thin wrapper to run a single Wombat analysis.
}
\details{
This function is just calling Wombat with a specified parameter
  parameter file.
}
\examples{
# generate a random pedigree structure for 20 offsprings, using 5 different
# sires and 7 different dams
set.seed(4)
animal_id = 1:20
sire_id = sample(1:5, size = 20, replace = T)
dam_id = sample(1:7, size = 20, replace = T)

# generate some random traits
weight = rnorm(n = 20, mean = 50)
length = rnorm(n = 20, mean = 100)

# assemble the information into one data frame
ped_pheno_data = data.frame(animal_id, sire_id, dam_id, weight, length)
ped_pheno_data

# save the files
wombat.write.pedigree(data = ped_pheno_data, animal.id = "animal_id",
  father.id = "sire_id", mother.id = "dam_id", convert.pedigree = T,
  file.path = "toto.ped")
wombat.write.data(data = ped_pheno_data, trait = "weight",
  animal.id = "animal_id", father.id = "sire_id", mother.id = "dam_id",
  write.mother = T, trait.factor = 2, convert.pedigree = T,
  filter.unknown.mother = T, file.path = "toto.dat")
wombat.write.parameters(trait = "weight", pedigree.file = "toto.ped",
  data.file = "toto.dat", var.residual = 2, var.animal = 2,
  var.mother = 2, maternal.effect = T, file.path = "toto.par")

# run the analysis
wombat.unit.run(parameter.file = "toto.par")
}

