% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wombat.write.parameters}
\alias{wombat.write.parameters}
\title{Generate a parameter file for an univariate Wombat analysis}
\usage{
wombat.write.parameters(trait, pedigree.file, data.file, var.residual,
  var.animal, var.mother = NA, maternal.effect = F, file.path)
}
\arguments{
\item{trait}{name of the trait to be analysed (\code{string}).}

\item{pedigree.file}{path to the pedigree file (e.g. generated by
\code{\link{wombat.write.pedigree}}).}

\item{data.file}{path to the data file (e.g. generated by
\code{\link{wombat.write.data}}).}

\item{var.residual}{numerical, initial value for residual variance.}

\item{var.animal}{numerical, initial value for animal variance.}

\item{var.mother}{numerical, initial value for maternal effect variance.}

\item{maternal.effect}{\code{Boolean}, should maternal effect be included
as a random factor?}

\item{file.path}{path to save the parameter file. This path should end with
  \code{".par"}.}
}
\value{
Nothing.
}
\description{
Produce a parameter file to be used for an univariate Wombat
  run.
}
\details{
IMPORTANT NOTE: the variance starting values used as arguments
  must be less than the variance of the trait (after the trait is
  multiplied by \code{trait.factor} in \code{\link{wombat.write.data}})-
  If this is not the case, Wombat will wait for a user input confirming
  that the run has to be made using large variance, and the R script will
  freeze waiting for this input.
}
\examples{
# generate a random pedigree structure for 20 offsprings, using 5 different
# sires and 7 different dams
set.seed(4)
animal_id = 1:20
sire_id = sample(1:5, size = 20, replace = T)
dam_id = sample(1:7, size = 20, replace = T)

# generate some random traits
weight = rnorm(n = 20, mean = 50)
length = rnorm(n = 20, mean = 100)

# assemble the information into one data frame
ped_pheno_data = data.frame(animal_id, sire_id, dam_id, weight, length)
ped_pheno_data

# save the files
wombat.write.pedigree(data = ped_pheno_data, animal.id = "animal_id",
  father.id = "sire_id", mother.id = "dam_id", convert.pedigree = T,
  file.path = "toto.ped")
wombat.write.data(data = ped_pheno_data, trait = "weight",
  animal.id = "animal_id", father.id = "sire_id", mother.id = "dam_id",
  write.mother = T, trait.factor = 2, convert.pedigree = T,
  filter.unknown.mother = T, file.path = "toto.dat")
wombat.write.parameters(trait = "weight", pedigree.file = "toto.ped",
  data.file = "toto.dat", var.residual = 2, var.animal = 2,
  var.mother = 2, maternal.effect = T, file.path = "toto.par")
}

